Split handling messages and set of functions
